Refactor(#18): 사용하지 않는 reservationId 삭제
