+++ src/main/java/com/widyu/pay/config/PaymentAuthInterceptor.java
@@ -1,5 +1,6 @@
 package com.widyu.pay.config;
 
+import com.widyu.pay.api.dto.PaymentProperties;
 import feign.RequestInterceptor;
 import feign.RequestTemplate;
 import java.nio.charset.StandardCharsets;
